move tests to src folder
