+++ src/tests/e2e/popup/disable.test.js
