+++ src/tests/e2e/popup/proxying/blocked.test.js
@@ -1,16 +1,26 @@
 import { getIconName, waitFor } from '../../utils'
 
+const settingsLink = `${global.extensionUrlPrefix}://${global.extensionId}/registry.html`
+
 describe('testing how processing blocked hostname affects UI', () => {
   let page
   let popUp
 
   beforeAll(async () => {
-    page = await global.getPage()
-    popUp = await global.getPopUp()
+    // configure region manually
+    const settingsPage = await global.getPage()
+
+    await settingsPage.goto(settingsLink)
+    await waitFor(1000)
+    await settingsPage.click('#select-toggle')
+    await settingsPage.click('.select-option[data-value="RU"]')
+    await settingsPage.close()
 
     // set timer to make sure that extension gets the registry
     await waitFor(10000)
 
+    page = await global.getPage()
+    popUp = await global.getPopUp()
     await page.goto('https://bbc.com', { timeout: 60000 })
     await waitFor(1000)
     await popUp.reload()
