+++ src/tests/e2e/popup/proxying/notBlocked.test.js
