+++ src/tests/e2e/popup/startup.test.js
